+++ app/scripts/controllers/index.js
@@ -42,3 +42,4 @@ export * from './channel-tabs/requestMatching'
 export * from './channel-tabs/userAccess'
 export * from './channel-tabs/routes'
 export * from './channel-tabs/dataControl'
+export * from './channel-tabs/alerts'
